Rename the components...
